Creating AddRetrospective modal, and styling related components
