+++ src/components/AddProject.vue
@@ -11,9 +11,9 @@
 		    	<div class="modal-content">
 						<div class="modal-header">
 							<h4 class="modal-title">Add a Project</h4>
-								<button type="button" class="close" data-dismiss="modal" aria-label="Fechar">
-									<span aria-hidden="true">&times;</span>
-								</button>
+							<button type="button" class="close" data-dismiss="modal" aria-label="Fechar">
+								<span aria-hidden="true">&times;</span>
+							</button>
 						</div>
 						<div class="modal-body">
 							<p><label>Name</label></p>
@@ -22,8 +22,8 @@
 							<input type = "text" v-model="description"></input><br>
 						</div>
 						<div class="modal-footer">
-							<button type="button" class="btn btn-primary" v-on:click="addProject" data-dismiss="modal">Save</button>
-							<button type="button" class="btn btn-secondary" data-dismiss="modal" >Close</button>
+							<button type="button" class="btn btn-primary falko-button" v-on:click="addProject" data-dismiss="modal">Save</button>
+							<button type="button" class="btn btn-secondary falko-button-grey" data-dismiss="modal" >Close</button>
 						</div>
 				</div>
 			</div>
